+++ core/src/com/myfyp/game/helper/StepCounterInterface.java
@@ -1,6 +1,5 @@
 package com.myfyp.game.helper;
 
 public interface StepCounterInterface {
-
-    public int getStepCount();
+    public float getStepCount();
 }
